move convertCTToPage to features/ct
